Refactor: Address TA feedback on XML metadata and precision
- Fixed copy-paste error in CurvatureAD XML description

- Updated all <float> parameters to <double> in XML files

- Added missing <constraints> for all numeric parameters

- Updated PixelType to double in C++ files to match XML precision

+++ src/CurvatureAD/CurvatureAD.cxx
@@ -8,7 +8,7 @@ int main(int argc, char * argv[])
 {
   PARSE_ARGS;
 
-  using PixelType = float;
+  using PixelType = double;
   constexpr unsigned int Dimension = 3;
   using ImageType = itk::Image<PixelType, Dimension>;
